Conflicts:
	sapling/maps/templates/maps/base.html
	sapling/maps/templates/maps/mapdata_diff.html
	sapling/maps/templates/maps/mapdata_list.html
	sapling/pages/models.py
	sapling/pages/static/js/ckeditor/_source/plugins/includepage/plugin.js
	sapling/pages/templates/pages/page_detail.html
	sapling/pages/templates/pages/page_versions.html
	sapling/pages/templates/pages/pagefile_diff.html
	sapling/pages/templatetags/pages_tags.py
	sapling/utils/views.py
	sapling/versionutils/diff/diffutils.py
	sapling/versionutils/diff/utils.py

+++ deb_utils/build_to_ppa.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
 UBUNTU_RELEASES="lucid maverick natty oneiric"
 
+if [[ "$1" = "" ]]
+then
+  echo "Usage: $0 <user/ppa>"
+  echo "Official LocalWiki PPA is localwiki/ppa"
+  echo "For testing use localwikidev/testing"
+  exit
+fi
+
+PPA=$1
+
 # Build orig.tar.gz
 cd ..
 rm -r deb_dist/
@@ -13,5 +23,5 @@ mv ${ORIG_FILE} /tmp
 cd ../deb_utils
 for DIST in $(echo ${UBUNTU_RELEASES}) ; do
     echo "Building and uploading for ${DIST} ..."
-    ./makedeb_ppa.sh "${DIST}" "/tmp/${ORIG_FILE}"
+    ./makedeb_ppa.sh "${DIST}" "/tmp/${ORIG_FILE}" "${PPA}"
 done
